Tests/LibWeb/Text/expected/navigation/ui-process-session-history-same-document.txt

after-push-state: location=?same-document-pushed
after-push-state: ui-current=?same-document-pushed
after-push-state: ui-relevant-entries=?same-document-replaced,?same-document-pushed*
after-push-state: ui-webcontent-known-relevant-entries=?same-document-replaced,?same-document-pushed*
after-push-state: ui-buttons=true,false
after-push-state: ui-uses-ui-step-coordinates=true
